requirements.txt
numpy~=1.16.6
pycococreatortools~=0.2.0
Pillow~=6.2.2
scikit-image~=0.14.5
matplotlib~=2.2.5
pycocotools~=2.0.2